Adding logging functionality and Admin role
Added login functionality with admin role to the application. Made changes to employee table by droping username and adding role to the employee.
Using one to one mapping we check for if the logged user is admin and return index of application.
Logging other user than admin gives error to the application.